+++ regions/core/core.py
@@ -3,20 +3,7 @@ import abc
 import copy
 import operator
 
-
-# Here we define global variables for the default `origin` and `mode` used
-# for WCS transformations throughout the `regions` package.
-#
-# Their purpose is to simplify achieving uniformity across the codebase.
-# They are mainly used as default arguments for methods that do WCS
-# transformations.
-#
-# They are private (with an underscore), not part of the public API, users
-# should not touch them.
-_DEFAULT_WCS_ORIGIN = 0
-_DEFAULT_WCS_MODE = 'all'
-
-VALID_MASK_MODES = {'center', 'exact', 'subpixels'}
+from .pixcoord import PixCoord
 
 __all__ = ['Region', 'PixelRegion', 'SkyRegion']
 
@@ -216,14 +203,14 @@ class PixelRegion(Region):
 
     @staticmethod
     def _validate_mode(mode, subpixels):
-        if mode not in VALID_MASK_MODES:
-            valid_modes = '/'.join(VALID_MASK_MODES)
-            raise ValueError(f"Invalid mask mode: {mode} (should be one "
-                             f"of {valid_modes}")
+        valid_modes = ('center', 'exact', 'subpixels')
+        if mode not in valid_modes:
+            raise ValueError(f'Invalid mask mode: {mode} (should be one '
+                             f'of {valid_modes}')
         if mode == 'subpixels':
             if not isinstance(subpixels, int) or subpixels <= 0:
-                raise ValueError(f"Invalid subpixels value: {subpixels} "
-                                 "(should be a strictly positive integer)")
+                raise ValueError(f'Invalid subpixels value: {subpixels} '
+                                 '(should be a strictly positive integer)')
 
     @abc.abstractmethod
     def as_artist(self, origin=(0, 0), **kwargs):
@@ -360,7 +347,6 @@ class SkyRegion(Region):
             The world coordinate system transformation to use to convert
             between sky and pixel coordinates.
         """
-        from .pixcoord import PixCoord
         pixel_region = self.to_pixel(wcs)
         pixcoord = PixCoord.from_sky(skycoord, wcs)
         return pixel_region.contains(pixcoord)