+++ regions/shapes/circle.py
@@ -9,8 +9,9 @@ from astropy.wcs.utils import pixel_to_skycoord
 from ..core import PixCoord, PixelRegion, SkyRegion, RegionMask, BoundingBox
 from .._utils.wcs_helpers import pixel_scale_angle_at_skycoord
 from .._geometry import circular_overlap_grid
+from ..core.metadata import RegionMeta, RegionVisual
 from ..core.attributes import (ScalarSky, ScalarPix, QuantityLength,
-                               ScalarLength, RegionVisual, RegionMeta)
+                               ScalarLength)
 
 __all__ = ['CirclePixelRegion', 'CircleSkyRegion']