+++ regions/shapes/point.py
@@ -3,7 +3,8 @@ import numpy as np
 from astropy.wcs.utils import pixel_to_skycoord, skycoord_to_pixel
 
 from ..core import PixCoord, PixelRegion, SkyRegion, BoundingBox
-from ..core.attributes import ScalarPix, ScalarSky, RegionMeta, RegionVisual
+from ..core.attributes import ScalarPix, ScalarSky
+from ..core.metadata import RegionMeta, RegionVisual
 
 __all__ = ['PointPixelRegion', 'PointSkyRegion']