+++ regions/shapes/polygon.py
@@ -6,7 +6,8 @@ from astropy.wcs.utils import skycoord_to_pixel, pixel_to_skycoord
 from ..core import PixelRegion, SkyRegion, RegionMask, BoundingBox, PixCoord
 from .._geometry import polygonal_overlap_grid
 from .._geometry.pnpoly import points_in_polygon
-from ..core.attributes import OneDPix, OneDSky, RegionMeta, RegionVisual
+from ..core.attributes import OneDPix, OneDSky
+from ..core.metadata import RegionMeta, RegionVisual
 
 __all__ = ['PolygonPixelRegion', 'PolygonSkyRegion']